Change defaults for VariableBaseMSM.
The default, user-friendly way is to multiply group elements by elements
in the scalar field. In an attempt to add some ergonomics,
I'm using the same template but for scalars and adding a big fat warning
that no checks is performed on the size.

The function that acts immediately on the bigints can also be called
explicitly if the caller knows that the same scalar vectors will be used
multiple times. Note that copying that vector only buys a few ms per
MSM of 2^20 elements or more.
The extra function `msm_bigint` also allows to perform  optimizations
on `msm` before the call to `msm_bigint`: instead of
normalizing into `BigInt`s with `u64` we can move to `i32` and use half the
number of buckets in Pippenger, exploiting group addition,
using the representation in buckets as [-w/2 + 1, w/2] instead of [0, w-1].

Use the frontend API `VariableBaseMSM::msm` in tests when possible.

Change (again) defaults for VariableBaseMSM not to panic.

After asking around, looks like most people assume that there is no bound
check (e.g. when committing to a vector or a polynomial). I'm therefore
reverting the change on the bound check