Fix test_sw_properties for some cofactor groups
The sw_properties test failed for groups with cofactors, for which the
first valid point (according to incrementing x) was on the prime order
subgroup.
The test assumed that the first point should have been of low order,
such that multiplying it with the cofactor put it in the high order
group.

This patch simplifies the code, and skips over these points.

Fixes #553