Set correct minimum supported Rust version
The minimum supported Rust version (MSRV) isn't correct in all crates,
this commit fixes it.

To find the minumum version [cargo msrv](https://crates.io/crates/cargo-msrv)
was used. Here are the results:

 - `ark-ff``: 1.63 due to dependency on `ark-serialize`
 - `ark-ff-asm`: 1.60 due to `ahash` 0.8.0
 - `arf-ff-macros`: 1.60 due to `ahash` 0.8.0
 - `ark-poly`: 1.63 due to dependency on `ark-serialize`
 - `ark-serialize`: 1.63 due to the use of [`core::array::from_fn`](https://doc.rust-lang.org/core/array/fn.from_fn.html)
 - `ark-serialize-derive`: 1.60 due to `ahash` 0.8.0
 - `ark-test-curves`: 1.63 due to depndency on `ark-ec`
 - `ark-algebra-test-templates`: 1.63 due to dependency on `ark-serialize`

In order to verify that now the set versions are at least the minimum, you
can run:

    find . -mindepth 2 -name Cargo.toml|parallel 'echo {} && cd `dirname {}` && cargo msrv verify'