feat: make AffineRepr::xy() more flexible
Currently [`AffineRepr::xy()` returns the field elements as references]. This is
only possible if the underlying data structure that implements that traits owns
those field elements and can directly have a reference to it.

Though there are cases, where providing those direct references is not possible.
For example if the underlying structure has a different layout, or if the fields
elements you're returning need to be wrapped in a newtype struct (that's my case).

With this change, `AffineRepr::xy()` now returns owned data. It's more flexible
as now the underlying data structure can have any shape as long as it can return
those field elements somehow.

[`AffineRepr::xy()` returns the field elements as references]: https://github.com/arkworks-rs/algebra/blob/4e4fc17e69178ab7f929602f17a5e769ebaf34c7/ec/src/lib.rs#L208